requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
gradio==6.4.0
gradio_client==2.0.3
greenlet==3.3.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.3
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
msgpack==1.1.2
namex==0.1.0
narwhals==2.15.0
numpy==2.4.1
opencv-python==4.13.0.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
packaging==26.0
pandas==2.3.3
pathlib==1.0.1
pillow==12.1.0
plotly==6.5.2
protobuf==5.29.5
psycopg2-binary==2.9.11
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safehttpx==0.1.7
scipy==1.17.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
starlette==0.50.0
streamlit==1.53.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf==1.1.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==2.0.1